Move streams to stream folder
