+++ src/stream/slpFolderStream.ts
