+++ src/utils/index.ts
@@ -1,7 +1,3 @@
 export * from "./dolphin";
 export * from "./slpFile";
-export * from "./slpStream";
-export * from "./slpWriter";
-export * from "./slpLiveStream";
-export * from "./slpFolderStream";
 export * from "./controls";
